requirements.txt
pyyaml==6.0.1
tortoise-orm==0.20.0
asyncpg==0.29.0
aerich==0.7.2
minio==7.2.3
boto3==1.34.14